bms/requirements.txt
Django==2.1.3
djangorestframework==3.11.0
Pillow==7.1.2
PyYAML==5.3.1
autopep8==1.5.3
PyJWT==1.7.1
